feat: погашения нал/карта у принявшего + дробное количество/единицы товара + фикс ложного офлайна (VPN/белые списки)
Три запроса владельца:

1. Рассрочка: способ оплаты погашения (нал/карта, мигр. 119) пишется и
   спрашивается при приёме; касса мастера видит погашения, принятые ИМ
   (created_by); движение денег/дашборд — разбивка «в т.ч. наличными/картой»;
   атрибуция по принявшему. Оборот от погашений не растёт (двойной счёт).

2. Товар не только в штуках: единицы шт/м/кг/л/уп/компл (чипсы в карточке),
   количество до 3 знаков (0.5 м шланга; мигр. 120: NUMERIC(12,3) + конверт
   legacy-юнитов + row_security off). QtyInput с черновиком (старый input
   съедал «0.5», на web мог тихо оставить строку ~0 ₽); вся арифметика
   количества через roundQty (float-грязь 1.3−1 ловилась бэкендом 400-кой).
   Услуги остаются целыми. Миграции прогнаны 3× на чистом PG16.

3. «Не грузит / нет сети» под VPN и в жёсткие окна: onlineManager больше не
   доверяет isInternetReachable (его проба ходила на clients3.google.com —
   зарезан операторами → ложный офлайн ставил ВСЕ запросы на паузу); пауза
   только при isConnected=false; NetInfo-проба переведена на ya.ru; captive
   закрыт HTML-стражами (probeHealth + нейтральные пробы баннера); таймауты
   подняты под VPN-латентность (12→15 с axios, 4→8/3→6 с пробы).

Ревью: 17 замечаний двумя волнами — все закрыты; web-ввод и миграции
проверены адверсариально (посимвольная симуляция + живой PG).
Дизайн: docs/BATCH_2026-07-05_UNITS_INSTALLMENT_CONNECTIVITY.md
Гейты: backend/frontend tc+lint+build, mobile tc+lint+jest 403/403.

+++ backend/src/employees/employees.service.ts
@@ -809,7 +809,8 @@ export class EmployeesService {
       topProducts: topProductRows.map((r) => ({
         productId: r.id as string,
         name: r.name as string,
-        count: parseInt(r.cnt) || 0,
+        // 120: количества дробные (NUMERIC(12,3)) — parseInt съедал бы "2.500".
+        count: parseFloat(r.cnt) || 0,
         // `photo` is optional in the contract — only emit it when present so the
         // payload stays clean for products without an image.
         ...(r.photo ? { photo: r.photo as string } : {}),
